Refactor: 코드 분리 및 주석 추가
